Restructure Components
