+++ app/javascript/components/Groceries/index.ts
@@ -1,0 +1,1 @@
+export { default as Groceries } from './Groceries';
