+++ app/javascript/components/Order/index.ts
@@ -1,0 +1,1 @@
+export { default as Order } from './Order'
