+++ app/javascript/components/index.ts
@@ -1,0 +1,3 @@
+export { Groceries } from './Groceries'
+export { Home } from './Home'
+export { Order } from './Order'
